Project Done

+++ modules/event/views.py
@@ -8,12 +8,12 @@ from flask_login import (LoginManager, current_user, login_manager,
                          login_required, login_user, logout_user)
 
 from app import db, login_manager
+from modules.attendance.models import Attendance
 from modules.intake.models import Semester
 from modules.user.models import User
 
+from . import event_bp
 from .models import Event
-
-event_bp = Blueprint('event', __name__, url_prefix='/event')
 
 bcrypt = Bcrypt()
 
@@ -31,22 +31,21 @@ def create():
         end_date = request.form['end_datetime']
         venue = request.form['venue']
         
-        validate_start_date(start_date)
         start_date = datetime.strptime(start_date, "%Y-%m-%dT%H:%M")
         if start_date < datetime.now():
-            flash('Start date can not be before today', 'danger')
+            flash('Start date cannot be before today', 'danger')
             return redirect(url_for('event.create'))
 
         end_date = datetime.strptime(end_date, "%Y-%m-%dT%H:%M")
         if end_date < start_date:
-            flash('End date can not be before start date', 'danger')
+            flash('End date cannot be before start date', 'danger')
             return redirect(url_for('event.create'))
 
         event = Event(title=title, description=description, start_date=start_date, end_date=end_date, venue=venue)
         db.session.add(event)
         db.session.commit()
 
-        flash('Event created successfully', 'success')
+        flash('Event created successfully!', 'success')
         return redirect(url_for('event.list'))
 
     return render_template('events/create.html')
@@ -55,30 +54,28 @@ def create():
 @event_bp.route('/list')
 @login_required
 def list():
-    events = Event.query.all()
+    events = Event.query.order_by(Event.start_date.desc()).all()
+    now = datetime.now()
     
-    events = [event for event in events if event.end_date > datetime.now()]
-    
-    total_events = len(events)
-    
-    return render_template('events/list.html', events=events, total_events=total_events)
+    return render_template('events/list.html', events=events, now=now)
 
-@event_bp.route('/show/<int:event_id>')
-@login_required
-def show(event_id):
-    event = Event.query.get(event_id)
-    return render_template('events/show.html', event=event)
 
-#delete event
-@event_bp.route('/delete/<int:event_id>')
+@event_bp.route('/delete/<int:event_id>', methods=['GET', 'POST'])
 @login_required
 def delete(event_id):
-    event = Event.query.get(event_id)
-    db.session.delete(event)
-    db.session.commit()
+    if request.method == 'POST':
+        event = Event.query.get(event_id)
+        attendances = Attendance.query.filter_by(event_id=event_id).all()
+        
+        for attendance in attendances:
+            db.session.delete(attendance)
+        
+        db.session.delete(event)
+        db.session.commit()
 
-    flash('Event deleted successfully', 'success')
-    return redirect(url_for('event.list'))
+        flash('Event deleted successfully!', 'success')
+        return redirect(url_for('event.list'))
+    
 
 def validate_start_date(start_date):
     start_date = datetime.strptime(start_date, "%Y-%m-%dT%H:%M")
